Trait core_ethereum_db::traits::HoprCoreEthereumDbActions

pub trait HoprCoreEthereumDbActions {

    // Required methods
    fn get_current_ticket_index<'life0, 'life1, 'async_trait>(
        &'life0 self,
        channel_id: &'life1 Hash
    ) -> Pin<Box<dyn Future<Output = Result<Option<U256>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn set_current_ticket_index<'life0, 'life1, 'async_trait>(
        &'life0 mut self,
        channel_id: &'life1 Hash,
        index: U256
    ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn increase_current_ticket_index<'life0, 'life1, 'async_trait>(
        &'life0 mut self,
        channel_id: &'life1 Hash
    ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn ensure_current_ticket_index_gte<'life0, 'life1, 'async_trait>(
        &'life0 mut self,
        channel_id: &'life1 Hash,
        index: U256
    ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn get_tickets<'life0, 'async_trait>(
        &'life0 self,
        signer: Option<Address>
    ) -> Pin<Box<dyn Future<Output = Result<Vec<Ticket>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_unrealized_balance<'life0, 'life1, 'async_trait>(
        &'life0 self,
        channel: &'life1 Hash
    ) -> Pin<Box<dyn Future<Output = Result<Balance>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn get_channel_epoch<'life0, 'life1, 'async_trait>(
        &'life0 self,
        channel: &'life1 Hash
    ) -> Pin<Box<dyn Future<Output = Result<Option<U256>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn cleanup_invalid_channel_tickets<'life0, 'life1, 'async_trait>(
        &'life0 mut self,
        channel: &'life1 ChannelEntry
    ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn mark_rejected<'life0, 'life1, 'async_trait>(
        &'life0 mut self,
        ticket: &'life1 Ticket
    ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn get_pending_acknowledgement<'life0, 'life1, 'async_trait>(
        &'life0 self,
        half_key_challenge: &'life1 HalfKeyChallenge
    ) -> Pin<Box<dyn Future<Output = Result<Option<PendingAcknowledgement>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn store_pending_acknowledgment<'life0, 'async_trait>(
        &'life0 mut self,
        half_key_challenge: HalfKeyChallenge,
        pending_acknowledgment: PendingAcknowledgement
    ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn replace_unack_with_ack<'life0, 'life1, 'async_trait>(
        &'life0 mut self,
        half_key_challenge: &'life1 HalfKeyChallenge,
        ack_ticket: AcknowledgedTicket
    ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn get_acknowledged_tickets<'life0, 'async_trait>(
        &'life0 self,
        filter: Option<ChannelEntry>
    ) -> Pin<Box<dyn Future<Output = Result<Vec<AcknowledgedTicket>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_acknowledged_tickets_count<'life0, 'async_trait>(
        &'life0 self,
        filter: Option<ChannelEntry>
    ) -> Pin<Box<dyn Future<Output = Result<usize>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn prepare_aggregatable_tickets<'life0, 'life1, 'async_trait>(
        &'life0 mut self,
        channel_id: &'life1 Hash,
        epoch: u32,
        index_start: u64,
        index_end: u64
    ) -> Pin<Box<dyn Future<Output = Result<Vec<AcknowledgedTicket>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn get_acknowledged_tickets_range<'life0, 'life1, 'async_trait>(
        &'life0 self,
        channel_id: &'life1 Hash,
        epoch: u32,
        index_start: u64,
        index_end: u64
    ) -> Pin<Box<dyn Future<Output = Result<Vec<AcknowledgedTicket>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn replace_acked_tickets_by_aggregated_ticket<'life0, 'async_trait>(
        &'life0 mut self,
        aggregated_ticket: AcknowledgedTicket
    ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_unacknowledged_tickets<'life0, 'async_trait>(
        &'life0 self,
        filter: Option<ChannelEntry>
    ) -> Pin<Box<dyn Future<Output = Result<Vec<UnacknowledgedTicket>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn update_acknowledged_ticket<'life0, 'life1, 'async_trait>(
        &'life0 mut self,
        ticket: &'life1 AcknowledgedTicket
    ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn get_packet_key<'life0, 'life1, 'async_trait>(
        &'life0 self,
        chain_key: &'life1 Address
    ) -> Pin<Box<dyn Future<Output = Result<Option<OffchainPublicKey>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn get_chain_key<'life0, 'life1, 'async_trait>(
        &'life0 self,
        packet_key: &'life1 OffchainPublicKey
    ) -> Pin<Box<dyn Future<Output = Result<Option<Address>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn link_chain_and_packet_keys<'life0, 'life1, 'life2, 'life3, 'async_trait>(
        &'life0 mut self,
        chain_key: &'life1 Address,
        packet_key: &'life2 OffchainPublicKey,
        snapshot: &'life3 Snapshot
    ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait,
             'life2: 'async_trait,
             'life3: 'async_trait;
    fn get_channel_to<'life0, 'life1, 'async_trait>(
        &'life0 self,
        dest: &'life1 Address
    ) -> Pin<Box<dyn Future<Output = Result<Option<ChannelEntry>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn get_channel_from<'life0, 'life1, 'async_trait>(
        &'life0 self,
        src: &'life1 Address
    ) -> Pin<Box<dyn Future<Output = Result<Option<ChannelEntry>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn update_channel_and_snapshot<'life0, 'life1, 'life2, 'life3, 'async_trait>(
        &'life0 mut self,
        channel_id: &'life1 Hash,
        channel: &'life2 ChannelEntry,
        snapshot: &'life3 Snapshot
    ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait,
             'life2: 'async_trait,
             'life3: 'async_trait;
    fn mark_acknowledged_tickets_neglected<'life0, 'async_trait>(
        &'life0 mut self,
        source: ChannelEntry
    ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_latest_block_number<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = Result<Option<u32>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn update_latest_block_number<'life0, 'async_trait>(
        &'life0 mut self,
        number: u32
    ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_latest_confirmed_snapshot<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = Result<Option<Snapshot>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_channel<'life0, 'life1, 'async_trait>(
        &'life0 self,
        channel: &'life1 Hash
    ) -> Pin<Box<dyn Future<Output = Result<Option<ChannelEntry>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn get_channels<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = Result<Vec<ChannelEntry>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_channels_open<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = Result<Vec<ChannelEntry>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_account<'life0, 'life1, 'async_trait>(
        &'life0 self,
        address: &'life1 Address
    ) -> Pin<Box<dyn Future<Output = Result<Option<AccountEntry>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn update_account_and_snapshot<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 mut self,
        account: &'life1 AccountEntry,
        snapshot: &'life2 Snapshot
    ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait,
             'life2: 'async_trait;
    fn get_accounts<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = Result<Vec<AccountEntry>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_public_node_accounts<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = Result<Vec<AccountEntry>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_redeemed_tickets_value<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = Result<Balance>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_redeemed_tickets_count<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = Result<usize>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_neglected_tickets_count<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = Result<usize>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_neglected_tickets_value<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = Result<Balance>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_losing_tickets_count<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = Result<usize>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn mark_redeemed<'life0, 'life1, 'async_trait>(
        &'life0 mut self,
        ticket: &'life1 AcknowledgedTicket
    ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn mark_losing_acked_ticket<'life0, 'life1, 'async_trait>(
        &'life0 mut self,
        ticket: &'life1 AcknowledgedTicket
    ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn get_rejected_tickets_value<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = Result<Balance>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_rejected_tickets_count<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = Result<usize>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_channel_x<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 self,
        src: &'life1 Address,
        dest: &'life2 Address
    ) -> Pin<Box<dyn Future<Output = Result<Option<ChannelEntry>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait,
             'life2: 'async_trait;
    fn get_channels_from<'life0, 'life1, 'async_trait>(
        &'life0 self,
        address: &'life1 Address
    ) -> Pin<Box<dyn Future<Output = Result<Vec<ChannelEntry>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn get_outgoing_channels<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = Result<Vec<ChannelEntry>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_channels_to<'life0, 'life1, 'async_trait>(
        &'life0 self,
        address: &'life1 Address
    ) -> Pin<Box<dyn Future<Output = Result<Vec<ChannelEntry>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn get_incoming_channels<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = Result<Vec<ChannelEntry>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_hopr_balance<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = Result<Balance>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn set_hopr_balance<'life0, 'life1, 'async_trait>(
        &'life0 mut self,
        balance: &'life1 Balance
    ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn get_ticket_price<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = Result<Option<U256>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn set_ticket_price<'life0, 'life1, 'async_trait>(
        &'life0 mut self,
        ticket_price: &'life1 U256
    ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn get_node_safe_registry_domain_separator<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = Result<Option<Hash>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn set_node_safe_registry_domain_separator<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 mut self,
        node_safe_registry_domain_separator: &'life1 Hash,
        snapshot: &'life2 Snapshot
    ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait,
             'life2: 'async_trait;
    fn get_channels_domain_separator<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = Result<Option<Hash>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn set_channels_domain_separator<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 mut self,
        channels_domain_separator: &'life1 Hash,
        snapshot: &'life2 Snapshot
    ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait,
             'life2: 'async_trait;
    fn get_channels_ledger_domain_separator<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = Result<Option<Hash>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn set_channels_ledger_domain_separator<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 mut self,
        channels_ledger_domain_separator: &'life1 Hash,
        snapshot: &'life2 Snapshot
    ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait,
             'life2: 'async_trait;
    fn add_hopr_balance<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 mut self,
        balance: &'life1 Balance,
        snapshot: &'life2 Snapshot
    ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait,
             'life2: 'async_trait;
    fn sub_hopr_balance<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 mut self,
        balance: &'life1 Balance,
        snapshot: &'life2 Snapshot
    ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait,
             'life2: 'async_trait;
    fn get_staking_safe_address<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = Result<Option<Address>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn set_staking_safe_address<'life0, 'life1, 'async_trait>(
        &'life0 mut self,
        safe_address: &'life1 Address
    ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn get_staking_module_address<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = Result<Option<Address>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn set_staking_module_address<'life0, 'life1, 'async_trait>(
        &'life0 mut self,
        module_address: &'life1 Address
    ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn get_staking_safe_allowance<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = Result<Balance>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn set_staking_safe_allowance<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 mut self,
        allowance: &'life1 Balance,
        snapshot: &'life2 Snapshot
    ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait,
             'life2: 'async_trait;
    fn is_network_registry_enabled<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = Result<bool>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn set_network_registry<'life0, 'life1, 'async_trait>(
        &'life0 mut self,
        enabled: bool,
        snapshot: &'life1 Snapshot
    ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn is_allowed_to_access_network<'life0, 'life1, 'async_trait>(
        &'life0 self,
        node: &'life1 Address
    ) -> Pin<Box<dyn Future<Output = Result<bool>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn set_allowed_to_access_network<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 mut self,
        node: &'life1 Address,
        allowed: bool,
        snapshot: &'life2 Snapshot
    ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait,
             'life2: 'async_trait;
    fn get_from_network_registry<'life0, 'life1, 'async_trait>(
        &'life0 self,
        stake_account: &'life1 Address
    ) -> Pin<Box<dyn Future<Output = Result<Vec<Address>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn is_eligible<'life0, 'life1, 'async_trait>(
        &'life0 self,
        account: &'life1 Address
    ) -> Pin<Box<dyn Future<Output = Result<bool>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn set_eligible<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 mut self,
        account: &'life1 Address,
        eligible: bool,
        snapshot: &'life2 Snapshot
    ) -> Pin<Box<dyn Future<Output = Result<Vec<Address>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait,
             'life2: 'async_trait;
    fn is_mfa_protected<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = Result<Option<Address>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn set_mfa_protected_and_update_snapshot<'life0, 'life1, 'async_trait>(
        &'life0 mut self,
        maybe_mfa_address: Option<Address>,
        snapshot: &'life1 Snapshot
    ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;

}

Required Methods

fn get_current_ticket_index<'life0, 'life1, 'async_trait>( &'life0 self, channel_id: &'life1 Hash ) -> Pin<Box<dyn Future<Output = Result<Option<U256>>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

fn set_current_ticket_index<'life0, 'life1, 'async_trait>( &'life0 mut self, channel_id: &'life1 Hash, index: U256 ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

fn increase_current_ticket_index<'life0, 'life1, 'async_trait>( &'life0 mut self, channel_id: &'life1 Hash ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

fn ensure_current_ticket_index_gte<'life0, 'life1, 'async_trait>( &'life0 mut self, channel_id: &'life1 Hash, index: U256 ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

fn get_tickets<'life0, 'async_trait>( &'life0 self, signer: Option<Address> ) -> Pin<Box<dyn Future<Output = Result<Vec<Ticket>>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

fn get_unrealized_balance<'life0, 'life1, 'async_trait>( &'life0 self, channel: &'life1 Hash ) -> Pin<Box<dyn Future<Output = Result<Balance>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

fn get_channel_epoch<'life0, 'life1, 'async_trait>( &'life0 self, channel: &'life1 Hash ) -> Pin<Box<dyn Future<Output = Result<Option<U256>>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

fn cleanup_invalid_channel_tickets<'life0, 'life1, 'async_trait>( &'life0 mut self, channel: &'life1 ChannelEntry ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

fn mark_rejected<'life0, 'life1, 'async_trait>( &'life0 mut self, ticket: &'life1 Ticket ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

fn get_pending_acknowledgement<'life0, 'life1, 'async_trait>( &'life0 self, half_key_challenge: &'life1 HalfKeyChallenge ) -> Pin<Box<dyn Future<Output = Result<Option<PendingAcknowledgement>>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

fn store_pending_acknowledgment<'life0, 'async_trait>( &'life0 mut self, half_key_challenge: HalfKeyChallenge, pending_acknowledgment: PendingAcknowledgement ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

fn replace_unack_with_ack<'life0, 'life1, 'async_trait>( &'life0 mut self, half_key_challenge: &'life1 HalfKeyChallenge, ack_ticket: AcknowledgedTicket ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

fn get_acknowledged_tickets<'life0, 'async_trait>( &'life0 self, filter: Option<ChannelEntry> ) -> Pin<Box<dyn Future<Output = Result<Vec<AcknowledgedTicket>>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Get all acknowledged tickets within the filter criteria.

fn get_acknowledged_tickets_count<'life0, 'async_trait>( &'life0 self, filter: Option<ChannelEntry> ) -> Pin<Box<dyn Future<Output = Result<usize>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Get count of acknowledged tickets within the filter criteria.

fn prepare_aggregatable_tickets<'life0, 'life1, 'async_trait>( &'life0 mut self, channel_id: &'life1 Hash, epoch: u32, index_start: u64, index_end: u64 ) -> Pin<Box<dyn Future<Output = Result<Vec<AcknowledgedTicket>>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Gets all acknowledged tickets in the channel and marks the as being aggregated

fn get_acknowledged_tickets_range<'life0, 'life1, 'async_trait>( &'life0 self, channel_id: &'life1 Hash, epoch: u32, index_start: u64, index_end: u64 ) -> Pin<Box<dyn Future<Output = Result<Vec<AcknowledgedTicket>>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

fn replace_acked_tickets_by_aggregated_ticket<'life0, 'async_trait>( &'life0 mut self, aggregated_ticket: AcknowledgedTicket ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

fn get_unacknowledged_tickets<'life0, 'async_trait>( &'life0 self, filter: Option<ChannelEntry> ) -> Pin<Box<dyn Future<Output = Result<Vec<UnacknowledgedTicket>>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Get all unacknowledged tickets within the filter criteria.

fn update_acknowledged_ticket<'life0, 'life1, 'async_trait>( &'life0 mut self, ticket: &'life1 AcknowledgedTicket ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

fn get_packet_key<'life0, 'life1, 'async_trait>( &'life0 self, chain_key: &'life1 Address ) -> Pin<Box<dyn Future<Output = Result<Option<OffchainPublicKey>>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

fn get_chain_key<'life0, 'life1, 'async_trait>( &'life0 self, packet_key: &'life1 OffchainPublicKey ) -> Pin<Box<dyn Future<Output = Result<Option<Address>>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

fn link_chain_and_packet_keys<'life0, 'life1, 'life2, 'life3, 'async_trait>( &'life0 mut self, chain_key: &'life1 Address, packet_key: &'life2 OffchainPublicKey, snapshot: &'life3 Snapshot ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait, 'life2: 'async_trait, 'life3: 'async_trait,

fn get_channel_to<'life0, 'life1, 'async_trait>( &'life0 self, dest: &'life1 Address ) -> Pin<Box<dyn Future<Output = Result<Option<ChannelEntry>>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

fn get_channel_from<'life0, 'life1, 'async_trait>( &'life0 self, src: &'life1 Address ) -> Pin<Box<dyn Future<Output = Result<Option<ChannelEntry>>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Get channel from peer with Ethereum address.

fn update_channel_and_snapshot<'life0, 'life1, 'life2, 'life3, 'async_trait>( &'life0 mut self, channel_id: &'life1 Hash, channel: &'life2 ChannelEntry, snapshot: &'life3 Snapshot ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait, 'life2: 'async_trait, 'life3: 'async_trait,

Update channel information.

fn mark_acknowledged_tickets_neglected<'life0, 'async_trait>( &'life0 mut self, source: ChannelEntry ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Delete acknowledged tickets belonging to a channel

fn get_latest_block_number<'life0, 'async_trait>( &'life0 self ) -> Pin<Box<dyn Future<Output = Result<Option<u32>>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Get the value of the lastest block number.

fn update_latest_block_number<'life0, 'async_trait>( &'life0 mut self, number: u32 ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Set the latest block number to this value.

fn get_latest_confirmed_snapshot<'life0, 'async_trait>( &'life0 self ) -> Pin<Box<dyn Future<Output = Result<Option<Snapshot>>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Get the latest confirmed snapshot.

fn get_channel<'life0, 'life1, 'async_trait>( &'life0 self, channel: &'life1 Hash ) -> Pin<Box<dyn Future<Output = Result<Option<ChannelEntry>>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Get channel from hash.

fn get_channels<'life0, 'async_trait>( &'life0 self ) -> Pin<Box<dyn Future<Output = Result<Vec<ChannelEntry>>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

TODO: unused?

fn get_channels_open<'life0, 'async_trait>( &'life0 self ) -> Pin<Box<dyn Future<Output = Result<Vec<ChannelEntry>>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Get all open channels.

fn get_account<'life0, 'life1, 'async_trait>( &'life0 self, address: &'life1 Address ) -> Pin<Box<dyn Future<Output = Result<Option<AccountEntry>>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Get account for address.

fn update_account_and_snapshot<'life0, 'life1, 'life2, 'async_trait>( &'life0 mut self, account: &'life1 AccountEntry, snapshot: &'life2 Snapshot ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait, 'life2: 'async_trait,

Update the account entry.

fn get_accounts<'life0, 'async_trait>( &'life0 self ) -> Pin<Box<dyn Future<Output = Result<Vec<AccountEntry>>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Get all accounts.

fn get_public_node_accounts<'life0, 'async_trait>( &'life0 self ) -> Pin<Box<dyn Future<Output = Result<Vec<AccountEntry>>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Get all public accounts.

fn get_redeemed_tickets_value<'life0, 'async_trait>( &'life0 self ) -> Pin<Box<dyn Future<Output = Result<Balance>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Get the total value of redeemed tickets.

fn get_redeemed_tickets_count<'life0, 'async_trait>( &'life0 self ) -> Pin<Box<dyn Future<Output = Result<usize>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Get the total number of redeemed tickets.

fn get_neglected_tickets_count<'life0, 'async_trait>( &'life0 self ) -> Pin<Box<dyn Future<Output = Result<usize>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Get the total number of neglected tickets.

fn get_neglected_tickets_value<'life0, 'async_trait>( &'life0 self ) -> Pin<Box<dyn Future<Output = Result<Balance>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Get the total value of neglected tickets.

fn get_losing_tickets_count<'life0, 'async_trait>( &'life0 self ) -> Pin<Box<dyn Future<Output = Result<usize>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Get the total number of losing tickets.

fn mark_redeemed<'life0, 'life1, 'async_trait>( &'life0 mut self, ticket: &'life1 AcknowledgedTicket ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Mark the ticket as redeemed.

fn mark_losing_acked_ticket<'life0, 'life1, 'async_trait>( &'life0 mut self, ticket: &'life1 AcknowledgedTicket ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Mark an acknowledged ticket as losing.

fn get_rejected_tickets_value<'life0, 'async_trait>( &'life0 self ) -> Pin<Box<dyn Future<Output = Result<Balance>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Get the total value of all rejected tickets.

fn get_rejected_tickets_count<'life0, 'async_trait>( &'life0 self ) -> Pin<Box<dyn Future<Output = Result<usize>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Get the count of all rejected tickets.

fn get_channel_x<'life0, 'life1, 'life2, 'async_trait>( &'life0 self, src: &'life1 Address, dest: &'life2 Address ) -> Pin<Box<dyn Future<Output = Result<Option<ChannelEntry>>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait, 'life2: 'async_trait,

Get channel from source to destination.

fn get_channels_from<'life0, 'life1, 'async_trait>( &'life0 self, address: &'life1 Address ) -> Pin<Box<dyn Future<Output = Result<Vec<ChannelEntry>>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Get all channels from a given address.

fn get_outgoing_channels<'life0, 'async_trait>( &'life0 self ) -> Pin<Box<dyn Future<Output = Result<Vec<ChannelEntry>>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Get all the outgoing channels from current node.

fn get_channels_to<'life0, 'life1, 'async_trait>( &'life0 self, address: &'life1 Address ) -> Pin<Box<dyn Future<Output = Result<Vec<ChannelEntry>>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Get all channels to a given address.

fn get_incoming_channels<'life0, 'async_trait>( &'life0 self ) -> Pin<Box<dyn Future<Output = Result<Vec<ChannelEntry>>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Get all the incoming channels from current node.

fn get_hopr_balance<'life0, 'async_trait>( &'life0 self ) -> Pin<Box<dyn Future<Output = Result<Balance>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Get the current balance.

fn set_hopr_balance<'life0, 'life1, 'async_trait>( &'life0 mut self, balance: &'life1 Balance ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Set balance as the current balance.

fn get_ticket_price<'life0, 'async_trait>( &'life0 self ) -> Pin<Box<dyn Future<Output = Result<Option<U256>>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Get the current ticket price.

fn set_ticket_price<'life0, 'life1, 'async_trait>( &'life0 mut self, ticket_price: &'life1 U256 ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Set new ticket price

fn get_node_safe_registry_domain_separator<'life0, 'async_trait>( &'life0 self ) -> Pin<Box<dyn Future<Output = Result<Option<Hash>>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Get the domain separator of node-safe-registry contract

fn set_node_safe_registry_domain_separator<'life0, 'life1, 'life2, 'async_trait>( &'life0 mut self, node_safe_registry_domain_separator: &'life1 Hash, snapshot: &'life2 Snapshot ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait, 'life2: 'async_trait,

Set the domain separator of node-safe-registry contract

fn get_channels_domain_separator<'life0, 'async_trait>( &'life0 self ) -> Pin<Box<dyn Future<Output = Result<Option<Hash>>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Get the domain separator of channels contract

fn set_channels_domain_separator<'life0, 'life1, 'life2, 'async_trait>( &'life0 mut self, channels_domain_separator: &'life1 Hash, snapshot: &'life2 Snapshot ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait, 'life2: 'async_trait,

Set the domain separator of channels contract

fn get_channels_ledger_domain_separator<'life0, 'async_trait>( &'life0 self ) -> Pin<Box<dyn Future<Output = Result<Option<Hash>>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Get the ledger domain separator of channels contract

fn set_channels_ledger_domain_separator<'life0, 'life1, 'life2, 'async_trait>( &'life0 mut self, channels_ledger_domain_separator: &'life1 Hash, snapshot: &'life2 Snapshot ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait, 'life2: 'async_trait,

Set the ledger domain separator of channels contract

fn add_hopr_balance<'life0, 'life1, 'life2, 'async_trait>( &'life0 mut self, balance: &'life1 Balance, snapshot: &'life2 Snapshot ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait, 'life2: 'async_trait,

Add balance to the current balance.

fn sub_hopr_balance<'life0, 'life1, 'life2, 'async_trait>( &'life0 mut self, balance: &'life1 Balance, snapshot: &'life2 Snapshot ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait, 'life2: 'async_trait,

Subtract balance from the current balance.

fn get_staking_safe_address<'life0, 'async_trait>( &'life0 self ) -> Pin<Box<dyn Future<Output = Result<Option<Address>>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Get the staking safe address

fn set_staking_safe_address<'life0, 'life1, 'async_trait>( &'life0 mut self, safe_address: &'life1 Address ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Sets the staking safe address

fn get_staking_module_address<'life0, 'async_trait>( &'life0 self ) -> Pin<Box<dyn Future<Output = Result<Option<Address>>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Get the staking module address

fn set_staking_module_address<'life0, 'life1, 'async_trait>( &'life0 mut self, module_address: &'life1 Address ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Sets the staking module address

fn get_staking_safe_allowance<'life0, 'async_trait>( &'life0 self ) -> Pin<Box<dyn Future<Output = Result<Balance>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Get the allowance for HoprChannels contract to transfer tokens on behalf of staking safe address

fn set_staking_safe_allowance<'life0, 'life1, 'life2, 'async_trait>( &'life0 mut self, allowance: &'life1 Balance, snapshot: &'life2 Snapshot ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait, 'life2: 'async_trait,

Sets the allowance for HoprChannels contract to transfer tokens on behalf of staking safe address

fn is_network_registry_enabled<'life0, 'async_trait>( &'life0 self ) -> Pin<Box<dyn Future<Output = Result<bool>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Check whether the Network Registry is enabled.

fn set_network_registry<'life0, 'life1, 'async_trait>( &'life0 mut self, enabled: bool, snapshot: &'life1 Snapshot ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Enable or disable network registry

fn is_allowed_to_access_network<'life0, 'life1, 'async_trait>( &'life0 self, node: &'life1 Address ) -> Pin<Box<dyn Future<Output = Result<bool>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Check whether node is allowed to participate in the network

fn set_allowed_to_access_network<'life0, 'life1, 'life2, 'async_trait>( &'life0 mut self, node: &'life1 Address, allowed: bool, snapshot: &'life2 Snapshot ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait, 'life2: 'async_trait,

Enable or disable access to network

fn get_from_network_registry<'life0, 'life1, 'async_trait>( &'life0 self, stake_account: &'life1 Address ) -> Pin<Box<dyn Future<Output = Result<Vec<Address>>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

fn is_eligible<'life0, 'life1, 'async_trait>( &'life0 self, account: &'life1 Address ) -> Pin<Box<dyn Future<Output = Result<bool>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Check if address as eligible to be operating in the network.

fn set_eligible<'life0, 'life1, 'life2, 'async_trait>( &'life0 mut self, account: &'life1 Address, eligible: bool, snapshot: &'life2 Snapshot ) -> Pin<Box<dyn Future<Output = Result<Vec<Address>>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait, 'life2: 'async_trait,

Set address as eligible to be operating in the network. returns affected node addresses

fn is_mfa_protected<'life0, 'async_trait>( &'life0 self ) -> Pin<Box<dyn Future<Output = Result<Option<Address>>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Check if account is protected by a MFA module (e.g. Gnosis Safe) returns MFA module address

fn set_mfa_protected_and_update_snapshot<'life0, 'life1, 'async_trait>( &'life0 mut self, maybe_mfa_address: Option<Address>, snapshot: &'life1 Snapshot ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Marks this account as being protected by a MFA module (e.g. Gnosis Safe) or removes it Some(Address) -> MFA present None -> no MFA